+++ Changelog.md
@@ -1,6 +1,7 @@
 ### 0.4.16 (unreleased)
 
 Features:
+ * ABI JSON: Include new field ``statemutability`` with values ``view``, ``nonpayable`` and ``payable``.
  * Parser: Display previous visibility specifier in error if multiple are found.
  * Syntax Checker: Support ``pragma experimental <feature>;`` to turn on experimental features.
  * Static Analyzer: Warn about large storage structures.